Create a method to create the plot for the inverse power method. Split up the methods to createPlot to make this easier. Ta'da.

+++ src/powermethod/MatrixGroup.java
@@ -90,7 +90,7 @@ public class MatrixGroup {
 	 * 
 	 * @return The inversePowerMethodIterations
 	 */
-	public double getInversePowerMethodIterations() {
+	public int getInversePowerMethodIterations() {
 		return inversePowerMethodIterations;
 	}
